# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: EDP Control Plane API
  description: |
    Enterprise data control plane for registering sources and datasets,
    authoring and deploying pipelines, enforcing data quality and governance
    policy, tracking lineage and impact, orchestrating runs, and recording an
    append-only audit trail.

    ## Conventions

    * All collection endpoints are paginated with `limit` and `offset`, and
      sorted with `sort_by` and `sort_dir`. See the shared parameters.
    * Errors always use the same envelope: `error.code`, `error.message`,
      `error.details` and a top-level `request_id` that matches the
      `X-Request-Id` response header.
    * Mutations are `PATCH` rather than `PUT`, so a caller updates only the
      fields it means to change.
    * Request bodies are validated before any state is written; a validation
      failure returns `422` with one entry per offending field.
  version: 0.1.0
  contact:
    name: Data Platform Engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://edp.internal/api/v1
    description: Internal deployment
security:
  - bearerAuth: []
...
paths:
...
  /policies:
    get:
      tags: [governance]
      operationId: listPolicies
      summary: List policies
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: policy_type
          in: query
          schema: { type: string }
        - name: resource_type
          in: query
          schema: { type: string }
        - name: resource_id
          in: query
          schema: { type: string, format: uuid }
        - name: effect
          in: query
          schema: { type: string, enum: [allow, deny, warn, require] }
        - name: enabled
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: A page of policies.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [governance]
      operationId: createPolicy
      summary: Create a policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PolicyWrite' }
      responses:
        '201':
          description: The created policy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Policy' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /policies/by-slug/{slug}:
    get:
      tags: [governance]
      operationId: getPolicyBySlug
      summary: Fetch a policy by slug
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: The policy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Policy' }
        '404': { $ref: '#/components/responses/NotFound' }
  /policies/retention:
    get:
      tags: [governance]
      operationId: getRetentionPolicy
      summary: Resolve the effective retention policy for a resource
      description: |
        Returns the most specific enabled retention policy for the resource. A
        policy scoped to the resource wins over one scoped to its type, which in
        turn wins over the global default. Returns `204` when nothing applies.
      parameters:
        - name: resource_type
          in: query
          required: true
          schema: { type: string }
        - name: resource_id
          in: query
          schema: { type: string, format: uuid }
        - name: dataset_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The effective retention policy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Policy' }
        '204': { description: No retention policy applies. }
        '400': { $ref: '#/components/responses/BadRequest' }
  /policies/evaluate:
    post:
      tags: [governance]
      operationId: evaluatePolicy
      summary: Evaluate governance policies against a request
      description: |
        Pure evaluation: no state is written and the request actor is irrelevant.
        All matching policies are returned so a caller can explain a decision,
        along with the effective decision after precedence is applied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_type, action]
              properties:
                resource_type: { type: string }
                resource_id: { type: string, format: uuid }
                dataset_id: { type: string, format: uuid }
                action: { type: string }
                classification: { type: string }
                actor_id: { type: string }
                actor_roles: { type: array, items: { type: string } }
                environment: { type: string, enum: [dev, staging, prod] }
                attributes:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: The evaluation result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyEvaluation' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /policies/{policyID}:
    parameters:
      - name: policyID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [governance]
      operationId: getPolicy
      summary: Fetch a policy
      responses:
        '200':
          description: The policy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Policy' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [governance]
      operationId: updatePolicy
      summary: Update a policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PolicyWritePartial' }
      responses:
        '200':
          description: The updated policy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Policy' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [governance]
      operationId: deletePolicy
      summary: Delete a policy
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
  /data-contracts:
    get:
      tags: [governance]
      operationId: listDataContracts
      summary: List data contracts
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: dataset_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string, enum: [draft, active, breached, retired] }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: A page of data contracts.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataContractPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [governance]
      operationId: createDataContract
      summary: Create a data contract
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DataContractWrite' }
      responses:
        '201':
          description: The created contract.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataContract' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /data-contracts/{contractID}:
    parameters:
      - name: contractID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [governance]
      operationId: getDataContract
      summary: Fetch a data contract
      responses:
        '200':
          description: The contract.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataContract' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [governance]
      operationId: updateDataContract
      summary: Update a data contract
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DataContractWritePartial' }
      responses:
        '200':
          description: The updated contract.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataContract' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [governance]
      operationId: deleteDataContract
      summary: Delete a data contract
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT verified against the configured JWKS. Issuer, audience and expiry are
        all checked. When authentication is disabled the API runs without any
        middleware and every request is attributed to a development actor.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Rows to skip.
      schema: { type: integer, minimum: 0, default: 0 }
    SortBy:
      name: sort_by
      in: query
      description: Column to sort on. Unknown columns are rejected.
      schema: { type: string }
    SortDir:
      name: sort_dir
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, maxLength: 100 }
...
  responses:
    BadRequest:
      description: The query string or path parameters were malformed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The request conflicts with current state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: The body failed validation. Every offending field is listed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code.
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - CONFLICT
                - VALIDATION_FAILED
                - RATE_LIMITED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
                - NOT_IMPLEMENTED
            message:
              type: string
              description: Human-readable summary. Safe to show a user.
            details:
              type: array
              description: Present for validation failures.
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
        request_id:
          type: string
          format: uuid
          description: Matches the X-Request-Id header, for correlating with logs.
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
...
    Policy:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        policy_type: { type: string }
        effect: { type: string, enum: [allow, deny, warn, require] }
        resource_type: { type: string }
        resource_id: { type: string, format: uuid, nullable: true }
        applies_to_classification: { type: string }
        applies_to_environment: { type: string }
        conditions: { type: object, additionalProperties: true }
        actions: { type: array, items: { type: string } }
        priority: { type: integer }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PolicyWrite:
      type: object
      required: [name, policy_type, effect]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        policy_type: { type: string }
        effect: { type: string, enum: [allow, deny, warn, require] }
        resource_type: { type: string }
        resource_id: { type: string, format: uuid }
        applies_to_classification: { type: string }
        applies_to_environment: { type: string }
        conditions: { type: object, additionalProperties: true }
        actions: { type: array, items: { type: string } }
        priority: { type: integer }
        enabled: { type: boolean }
    PolicyWritePartial:
      allOf:
        - $ref: '#/components/schemas/PolicyWrite'
    PolicyEvaluation:
      type: object
      properties:
        decision: { type: string, enum: [allow, deny, warn, require] }
        reason: { type: string }
        matched_policies:
          type: array
          items: { $ref: '#/components/schemas/Policy' }
    DataContract:
      type: object
      properties:
        id: { type: string, format: uuid }
        dataset_id: { type: string, format: uuid }
        version: { type: string }
        status: { type: string, enum: [draft, active, breached, retired] }
        schema_definition: { type: object, additionalProperties: true }
        sla: { type: object, additionalProperties: true }
        producer: { type: string }
        consumer: { type: string }
        signatures: { type: array, items: { $ref: '#/components/schemas/ContractSignature' } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ContractSignature:
      type: object
      properties:
        party: { type: string, enum: [producer, consumer] }
        signed_by: { type: string }
        signed_at: { type: string, format: date-time }
        signature_ref: { type: string }
    DataContractWrite:
      type: object
      required: [dataset_id, version]
      properties:
        dataset_id: { type: string, format: uuid }
        version: { type: string }
        schema_definition: { type: object, additionalProperties: true }
        sla: { type: object, additionalProperties: true }
        producer: { type: string }
        consumer: { type: string }
    DataContractWritePartial:
      allOf:
        - $ref: '#/components/schemas/DataContractWrite'
...
    PolicyPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/Policy' } }
    DataContractPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/DataContract' } }
...
    Page:
      type: object
      description: Every list endpoint returns the same envelope.
      required: [items, meta]
      properties:
        items:
          type: array
          items: {}
        meta: { $ref: '#/components/schemas/PageMeta' }